+++ src/components/app/app.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 
-import AppHeader from './app-header';
-import TodoList from './todo-list';
-import SearchPanel from './search-panel';
-import ItemStatusFilter from './item-status-filter';
-import ItemAddForm from './item-add-form';
+import AppHeader from '../app-header';
+import TodoList from '../todo-list';
+import SearchPanel from '../search-panel';
+import ItemStatusFilter from '../item-status-filter';
+import ItemAddForm from '../item-add-form';
 
 import './app.css';
