+++ src/components/todo-list-item/todo-list-item.js
@@ -21,17 +21,19 @@ const TodoListItem = ({ important, done,
         className="todo-list-item-label"
         onClick={onToggleDone}>{label}</span>
 
-      <button type="button"
-              className="btn btn-outline-success btn-sm float-right"
-              onClick={onToggleImportant}>
-        <i className="fa fa-exclamation"></i>
-      </button>
-
-      <button type="button"
+        <button type="button"
               className="btn btn-outline-danger btn-sm float-right"
               onClick={onDelete}>
         <i className="fa fa-trash-o"></i>
-      </button>
+        </button>
+        
+        <button type="button"
+              className="btn btn-outline-success btn-sm float-right"
+              onClick={onToggleImportant}>
+        <i className="fa fa-exclamation"></i>
+       </button>
+
+      
     </span>
   );
 };
