--- src/index.css
@@ -1,5 +1,0 @@
-.index {
-    position: absolute;
-    top: 20%;
-    left: 40%;
-}
