+++ src/index.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './components/app';
-import AppHeader from "./components/app-header";
 
 ReactDOM.render(<App />,
   document.getElementById('root'));
